backend/requirements.txt
requests
pandas
django
djangorestframework
django-cors-headers
yfinance
scikit-learn
xgboost
